requirements.txt
transformers
faiss-cpu
sentence-transformers
numpy
PyPDF2
streamlit
newspaper3k
lxml[html_clean]
